Update main.tf
The terraform template is missing the backend-lb load balancing rule. Without this rule you are unable to pass traffic to the active GW>

Suggested code worked but is giving errors due to two elements being set on for "health_prob_port" line 198. I tired to specific the backend-ln on line 219 but failed. So the deployment still works with an error at the end. It is none impacting.